+++ src/app/components/thank-you/thank-you.component.scss
