+++ src/app/modules/login/login.component.html
@@ -12,7 +12,7 @@
             <dynamic-ng-bootstrap-form [group]="formGroup" [layout]="formLayout" [model]="formModel">
               <ng-template modelId="password">
                 <div class="mt-2">
-                  <a href>Forgot your password?</a>
+                  <a routerLink='/reset-password'>Forgot your password?</a>
                 </div>
               </ng-template>
             </dynamic-ng-bootstrap-form>
