--- src/app/modules/registration/components/thank-you/thank-you.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'vc-thank-you',
-  templateUrl: './thank-you.component.html',
-  styleUrls: [
-    './thank-you.component.scss',
-  ],
-})
-export class ThankYouComponent {
-}
