--- src/app/modules/reset-password/dispatch-confirmation/dispatch-confirmation.component.html
@@ -1,1 +1,0 @@
-<p>dispatch-confirmation works!</p>
