--- src/app/modules/reset-password/dispatch-confirmation/dispatch-confirmation.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'vc-dispatch-confirmation',
-  templateUrl: './dispatch-confirmation.component.html',
-  styleUrls: [
-    './dispatch-confirmation.component.scss',
-  ],
-})
-export class DispatchConfirmationComponent {}
