+++ src/app/modules/reset-password/reset-password-form/reset-password-form.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'vc-reset-password',
+  templateUrl: './reset-password-form.component.html',
+  styleUrls: [
+    './reset-password-form.component.scss',
+  ],
+})
+export class ResetPasswordFormComponent {}
