+++ src/app/modules/reset-password/reset-password-routing.module.ts
@@ -1,11 +1,16 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { ThankYouComponent } from 'src/app/components/thank-you/thank-you.component';
 import { EmailRequestComponent } from './email-request/email-request.component';
 
 const routes: Routes = [
   {
     path: '',
     component: EmailRequestComponent,
+  },
+  {
+    path: 'thank-you',
+    component: ThankYouComponent,
   },
 ];
