--- src/app/modules/reset-password/reset-password/reset-password.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'vc-reset-password',
-  templateUrl: './reset-password.component.html',
-  styleUrls: [
-    './reset-password.component.scss',
-  ],
-})
-export class ResetPasswordComponent {}
